refactor: split reader_base composition and ontology DC-circuit solver
Extract compose_property_mapping/narrow_property_mapping/_compose_family_surface
from reader_base.py into a new composition.py concern module, mirroring the
existing class_guards/credentials split.

Extract the shared DC-circuit Laplacian/Dirichlet solver primitives from
semantic_field.py into a new dc_solver.py, and point discovery.py at the
public dc_solver functions instead of importing semantic_field's
underscore-prefixed internals across module boundaries.

Split ontology/registry.py: move the NamespaceOntology/RelationshipRule
dataclasses and YAML parser into a new models.py, leaving OntologyRegistry
as a pure cache/lookup class.

No behavior change; all three are pure move/split refactors.
