+++ open_kgo/feature_groups/kg/ontology/registry.py
@@ -8,7 +8,9 @@ Loads a YAML ontology definition file and provides lookups for:
 The registry is keyed by namespace so rules are reusable across any connector
 or dataset that operates in the same domain (e.g. every movie KG shares the
 ``movie`` namespace ontology regardless of whether it lives in Neo4j, NetworkX,
-or RDF).
+or RDF). The parsed ontology shape (``NamespaceOntology`` / ``RelationshipRule``)
+and the YAML parser live in the sibling ``models`` module; this module owns the
+process-wide cache and namespace/path lookup.
 
 Usage::
 
@@ -20,47 +22,10 @@ Usage::
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, ClassVar
 
-
-@dataclass(frozen=True)
-class RelationshipRule:
-    """Domain and range type constraints for one relationship type."""
-
-    domain: str
-    range_type: str
-    weight: float = 1.0
-
-
-@dataclass
-class NamespaceOntology:
-    """All ontology rules for one namespace / domain."""
-
-    namespace: str
-    entity_valid_outgoing: dict[str, frozenset[str]]
-    relationships: dict[str, RelationshipRule]
-
-    def is_valid_edge(self, entity_type: str, relationship: str) -> bool:
-        """Return True if ``relationship`` is a valid outgoing edge from ``entity_type``.
-
-        Unknown entity types pass through (True) so connectors that carry
-        entities not declared in the ontology are not silently broken.
-        """
-        valid = self.entity_valid_outgoing.get(entity_type)
-        if valid is None:
-            return True
-        return relationship in valid
-
-    def get_range_type(self, relationship: str) -> str | None:
-        """Return the entity type at the far end of ``relationship``, or None."""
-        rule = self.relationships.get(relationship)
-        return rule.range_type if rule else None
-
-    def valid_next_hops(self, entity_type: str) -> frozenset[str]:
-        """Return the set of valid outgoing relationship types for ``entity_type``."""
-        return self.entity_valid_outgoing.get(entity_type, frozenset())
+from open_kgo.feature_groups.kg.ontology.models import NamespaceOntology, parse_ontology
 
 
 class OntologyRegistry:
@@ -118,42 +83,8 @@ class OntologyRegistry:
 
     @classmethod
     def _parse(cls, data: Any) -> NamespaceOntology:
-        """Parse a loaded YAML mapping into a ``NamespaceOntology``.
-
-        Raises ``ValueError`` (the same error type ``load_file`` already
-        documents for duplicate namespaces) on a malformed file — a non-mapping
-        top level, a missing ``namespace``, or a relationship that omits its
-        ``domain`` / ``range`` — so a bad ontology surfaces a clear, typed
-        error rather than a raw ``KeyError`` / ``TypeError`` from indexing.
-        """
-        if not isinstance(data, dict):
-            raise ValueError(f"ontology file must contain a YAML mapping at the top level, got {type(data).__name__}.")
-        if "namespace" not in data:
-            raise ValueError("ontology file is missing the required 'namespace' key.")
-        namespace = str(data["namespace"])
-
-        entity_valid_outgoing: dict[str, frozenset[str]] = {}
-        for name, spec in (data.get("entities") or {}).items():
-            valid_outgoing = spec.get("valid_outgoing") if isinstance(spec, dict) else None
-            entity_valid_outgoing[str(name)] = frozenset(valid_outgoing or [])
-
-        relationships: dict[str, RelationshipRule] = {}
-        for name, spec in (data.get("relationships") or {}).items():
-            if not isinstance(spec, dict) or "domain" not in spec or "range" not in spec:
-                raise ValueError(
-                    f"ontology relationship {name!r} must be a mapping declaring both 'domain' and 'range'."
-                )
-            relationships[str(name)] = RelationshipRule(
-                domain=str(spec["domain"]),
-                range_type=str(spec["range"]),
-                weight=float(spec.get("weight", 1.0)),
-            )
-
-        return NamespaceOntology(
-            namespace=namespace,
-            entity_valid_outgoing=entity_valid_outgoing,
-            relationships=relationships,
-        )
+        """Parse a loaded YAML mapping into a ``NamespaceOntology``; see ``models.parse_ontology``."""
+        return parse_ontology(data)
 
     @classmethod
     def get(cls, namespace: str) -> NamespaceOntology | None:
